FEAT: add GetReceivedEvent overload with timeout for EventConsumerHost
We already have a 'GetReceivedEvent' method with a retry count as parameter to define more reliable assertions; but after consideration this would benefit to have a overload that takes a timeout instead of a retry count.

With this, the tester doesn't have to worry about internal retry cylces but can instead think about what a reasonable time range would be for the test.

#52